LayoutTests/fast/js/char-at-expected.txt

"", omitted: "", NaN
"", undefined: "", NaN
"", 0: "", NaN
"", null: "", NaN
"", false: "", NaN
"", true: "", NaN
"", 0: "", NaN
"", 0.1: "", NaN
"", 999: "", NaN
"", Infinity: "", NaN
"", -1: "", NaN
"", -Infinity: "", NaN
"", NaN: "", NaN
"x", omitted: "x", 120
"x", undefined: "x", 120
"x", 0: "x", 120
"x", null: "x", 120
"x", false: "x", 120
"x", true: "", NaN
"x", 0: "x", 120
"x", 0.1: "x", 120
"x", 999: "", NaN
"x", Infinity: "", NaN
"x", -1: "", NaN
"x", -Infinity: "", NaN
"x", NaN: "", NaN
"xy", omitted: "x", 120
"xy", undefined: "x", 120
"xy", 0: "x", 120
"xy", null: "x", 120
"xy", false: "x", 120
"xy", true: "y", 121
"xy", 0: "x", 120
"xy", 0.1: "x", 120
"xy", 999: "", NaN
"xy", Infinity: "", NaN
"xy", -1: "", NaN
"xy", -Infinity: "", NaN
"xy", NaN: "", NaN
